src/main/java/master/springframework/usermanagement/controllers/IndexController.java
package master.springframework.usermanagement.controllers;

import lombok.extern.slf4j.Slf4j;
import master.springframework.usermanagement.services.UserService;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

@Controller
@Slf4j
public class IndexController {

    private final UserService userService;

    public IndexController(UserService userService) {
        this.userService = userService;
    }

    @RequestMapping({"", "/", "/index"})
    public String getIndexPage(Model model) {
        log.debug("Getting Index page");

        model.addAttribute("users", userService.getUsers());

        return "index";
    }
}
